server/requirements.txt
aniso8601==1.1.0
base58==0.2.3
boto3==1.4.0
botocore==1.4.56
click==6.6
docutils==0.12
Flask==0.11.1
Flask-RESTful==0.3.5
futures==3.0.5
hjson==2.0.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kappa==0.6.0
lambda-packages==0.8.0
MarkupSafe==0.23
placebo==0.8.1
pymongo==3.3.0
python-dateutil==2.5.3
python-slugify==1.2.1
pytz==2016.7
PyYAML==3.12
requests==2.11.1
s3transfer==0.1.7
six==1.10.0
tqdm==4.8.4
troposphere==1.8.1
Unidecode==0.4.19
Werkzeug==0.11.11
wsgi-request-logger==0.4.5
zappa==0.26.0
